# Improve yaml duplication checker performance